+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `iter_move_on_after()` and `iter_fail_after()` are iterator wrappers
   that apply a timeout to a single iteration.
+- `AsyncValue.eventual_values()` supports a `held_for` option.
 
 ## trio-util 0.6.0 (2021-09-05)
 ### Added